Patrio frontend
this is so you can have a feel of their demand they want their website to have a red fill and really have a design feel. But if you have a better idea I'm up for it. but first please finish the backend. Also b4 u start try to design a flowchart to guide u with Microsoft word or any other better option.